[refactor] 볼드 처리 메서드 UILabel extension에서 Base UI 요소들로 이동 #263
