+++ Happiggy-bank/Happiggy-bank/Base/UI/View/BaseLabel.swift
@@ -12,8 +12,8 @@ import UIKit
 final class BaseLabel: UILabel {
 
     // MARK: - Properties
-
-    private let fontManager: FontPublishing = FontManager.shared
+    
+    private let fontManager: FontManager = FontManager.shared
     private var cancellable: AnyCancellable?
 
 
@@ -32,16 +32,24 @@ final class BaseLabel: UILabel {
     }
 
 
-    // MARK: - Functions
+    // MARK: - Attribute Modifying Functions
+
+    /// 강조 처리
+    func bold() {
+        self.updateFont(to: self.fontManager.font, isBold: true)
+    }
+    
+
+    // MARK: - Configuration Functions
 
     private func subscribeToFontPublisher() {
         self.cancellable = fontManager.fontPublisher
             .receive(on: DispatchQueue.main)
-            .sink { [weak self] in self?.updateFont(to: $0) }
+            .sink { [weak self] in self?.updateFont(to: $0, isBold: self?.font.isBold == true) }
     }
 
-    private func updateFont(to newFont: CustomFont) {
-        let fontName = self.font.isBold ? newFont.bold : newFont.regular
+    private func updateFont(to newFont: CustomFont, isBold: Bool) {
+        let fontName = isBold ? newFont.bold : newFont.regular
         self.font = UIFont(name: fontName, size: self.font.pointSize)
     }
 }
